Rework the searchable CDT implementation to use more intuitive system for comparisons
As a side effect of changing the internal comparison system, some of the external
functions could be changed to be more clear and leak less internal details into
the interface

+++ include/arch/arm/arch/object/capspace.h
@@ -11,7 +11,7 @@
 #ifndef __ARCH_OBJECT_CAPSPACE_H
 #define __ARCH_OBJECT_CAPSPACE_H
 
-#define FMAPPED_OBJECT_HIGH(a) ((a) >> 12)
+#define FMAPPED_OBJECT_HIGH(a) ((uint32_t)(a) >> 12)
 
 enum capSpaceType {
     capSpaceUntypedMemory,
